requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==2.2.1
streamlit==1.41.1
psycopg2-binary==2.9.10